+++ skillctl/forensics/__init__.py
@@ -1,0 +1,5 @@
+"""Incident forensics for SkillsOps (Milestone 4)."""
+
+from skillctl.forensics.query import ForensicQuery
+
+__all__ = ["ForensicQuery"]